Created the game function that puts the program together
- created the function "game" that uses a for loop to call on "playRound" 5 times. The function prompts the user to enter a choice between rock paper and scissors for each round and calls on the "getComputerChoice" function to randomly determine the computers choice for each round.

if the player count is equal to 3 and the computer count less that 3, the "detectWinner" function is called displays the winner. Vice versa is the computers count was equal to 3.

If the both players points are below three (meaning a full set of 5 rounds has been played), then the "detectWinner" function is called on again to display the winner of the round